delete transactions db
